requirements.txt
groq
gtts
pygame
SpeechRecognition
python-dotenv
pydub